+++ R/utils.R
@@ -11,7 +11,7 @@ globalVariables(c(".", ":="))
 #' Any values between brackets {} will be replaced with `dunlin::render_safe`.
 #' @export
 var_labels_for <- function(df, vars) {
-  checkmate::assert_names(colnames(df), must.include = vars, what = "colnames")
+  assert_names(colnames(df), must.include = vars, what = "colnames")
   render_safe(unname(formatters::var_labels(df, fill = TRUE)[vars]))
 }
 
@@ -50,7 +50,7 @@ smart_prune <- function(tlg) {
 #' @keywords internal
 #'
 std_postprocess <- function(tlg, ind = 2L, ...) {
-  checkmate::assert_int(ind, lower = 0L)
+  assert_int(ind, lower = 0L)
 
   res <- report_null(tlg)
   table_inset(res) <- ind
@@ -75,8 +75,8 @@ std_postprocess <- function(tlg, ind = 2L, ...) {
 #' fun(c(123, 567.89))
 #'
 h_format_dec <- function(digits, format, ne = FALSE) {
-  checkmate::assert_integerish(digits, lower = 0)
-  checkmate::assert_string(format)
+  assert_integerish(digits, lower = 0)
+  assert_string(format)
   if (any(is.na(digits))) {
     function(x, ...) {
       ""
@@ -117,7 +117,7 @@ fuse_sequentially <- function(x, y) {
 #' @export
 grob_list <- function(...) {
   ret <- list(...)
-  checkmate::assert_list(ret, types = c("grob"))
+  assert_list(ret, types = c("grob"))
   structure(
     ret,
     class = c("grob_list", "list")
@@ -129,7 +129,7 @@ grob_list <- function(...) {
 #' @export
 gg_list <- function(...) {
   ret <- list(...)
-  checkmate::assert_list(ret, types = c("ggplot"))
+  assert_list(ret, types = c("ggplot"))
   structure(
     ret,
     class = c("gg_list", "list")
@@ -164,7 +164,7 @@ lvls.factor <- function(x) {
 
 #' @keywords internal
 quote_str <- function(x) {
-  checkmate::assert_string(x)
+  assert_string(x)
   paste0("`", x, "`")
 }
 
@@ -199,4 +199,12 @@ do_call <- function(what, args) {
   new_args <- lapply(arg_names, as.symbol)
   names(new_args) <- names(args)
   do.call(what, new_args, envir = args_env)
+}
+
+#' Modify character
+#' @keywords internal
+modify_character <- function(x, y) {
+  assert_character(x, names = "unique", null.ok = TRUE)
+  assert_character(y, names = "unique", null.ok = TRUE)
+  c(y, x)[unique(c(names(y), names(x)))]
 }