+++ tests/testthat/test-kmg01.R
@@ -4,7 +4,7 @@ test_that("kmg01 works as expected", {
   filter_data <- dunlin::log_filter(syn_data, PARAMCD == "OS", "adtte")
   pre_data <- expect_silent(kmg01_pre(filter_data, dataset = "adtte"))
   raw_res <- expect_silent(kmg01_main(pre_data, dataset = "adtte"))
-  checkmate::assert_true(grid::is.grob(raw_res))
+  expect_true(grid::is.grob(raw_res))
 })
 
 # kmg01 ----
@@ -18,9 +18,9 @@ test_that("kmg01 works as expected with custom color set", {
 
   filter_data <- dunlin::log_filter(syn_data, PARAMCD == "OS", "adtte")
   res <- expect_silent(run(kmg01, filter_data, dataset = "adtte", col = col))
-  checkmate::assert_true(grid::is.grob(res))
+  expect_true(grid::is.grob(res))
   res <- expect_silent(run(kmg01, filter_data, dataset = "adtte", col = unname(col)))
-  checkmate::assert_true(grid::is.grob(res))
+  expect_true(grid::is.grob(res))
 })
 
 test_that("kmg01 works if change pvalue, ties and conf level", {
@@ -31,7 +31,7 @@ test_that("kmg01 works if change pvalue, ties and conf level", {
     ties = "efron",
     conf_level = 0.99
   ))
-  checkmate::assert_true(grid::is.grob(res))
+  expect_true(grid::is.grob(res))
 })
 
 
@@ -41,5 +41,16 @@ test_that("kmg01 works if change annotation position", {
     dataset = "adtte", annot_surv_med = FALSE,
     position_coxph = c(0.4, 0.5), position_surv_med = c(1, 0.7)
   ))
-  checkmate::assert_true(grid::is.grob(res))
+  expect_true(grid::is.grob(res))
+})
+
+
+test_that("kmg01 works for stratified anlaysis", {
+  filter_data <- dunlin::log_filter(syn_data, PARAMCD == "OS", "adtte")
+  res <- expect_silent(run(kmg01, filter_data,
+    dataset = "adtte", annot_surv_med = FALSE,
+    position_coxph = c(0.4, 0.5), position_surv_med = c(1, 0.7),
+    strat = c("STRATA1", "STRATA2")
+  ))
+  expect_true(grid::is.grob(res))
 })